+++ WGT-BankingApplication/StandingOrder.cs
@@ -7,7 +7,7 @@ class StandingOrder : PersonalAccount
     private DateTime _date;
     private string _reference;
 
-    public StandingOrder(Customer Customer, int InitialDeposit, string Payee, decimal Amount, DateTime Date, string Reference) : base(Customer, InitialDeposit)
+    public StandingOrder(Customer Customer, decimal InitialDeposit, string Payee, decimal Amount, DateTime Date, string Reference) : base(Customer, InitialDeposit)
     {
         _payee = Payee;
         _amount = Amount;
